Rename test files to follow the same conventions as other test modules
